+++ Source/WinCopies.Collections.Shared/DotNetFix/Generic/LinkedCollections/LinkedListNodeEnumerator.cs
@@ -15,6 +15,9 @@
  * You should have received a copy of the GNU General Public License
  * along with the WinCopies Framework.  If not, see <https://www.gnu.org/licenses/>. */
 
+#if WinCopies2
+using System;
+#endif
 using System.Collections;
 using System.Collections.Generic;